Solved day 02 part 01 after giving SPL long powers

+++ compiler/spl-1.2.1/telma.h
@@ -27,7 +27,7 @@ USA.
    (telma.h? Read it backwards!)
 */
 
-int yyerror(char *s);
+long yyerror(char *s);
 int yylex(void);
 
 extern int yylineno;
